feat!: drop custom headers support and share session between auth and client
Remove the direct support for custom headers
(the session can be created and headers added),
and ensure the same session is shared between
the connection and the client.
This is important to support custom headers
(e.g. keycloak behind CloudFlare). In this case,
headers must also be present for the auth requests.