feat: optionally return the raw http response
Keycloak's current API is weird and doesn't return the created item in
most post requests. The only way to actually get the ID of the created
resource is to look at the headers `location` attribute.

This commit adds a way to get the raw response from a call, instead of
the parsed body. Just pass `return_raw_response=True` to the
get/post/... call.